% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{gpt2}
\alias{gpt2}
\title{Sample from a model conditioning on input}
\usage{
gpt2(prompt = "Hello my name is", model = c("124M", "345M", "774M"),
  length = NULL, temperature = 1, top_k = 0)
}
\arguments{
\item{prompt}{Input string to condition on.}

\item{model}{one of `c("124M", "345M", "774M")` (default is "124M")}

\item{length}{Number of tokens in generated text, if NULL (default), is
determined by model hyperparameters}

\item{temperature}{Float value controlling randomness in boltzmann
distribution. Lower temperature results in less random completions. As the
temperature approaches zero, the model will become deterministic and
repetitive. Higher temperature results in more random completions.
Default: 1.}

\item{top_k}{Integer value controlling diversity. 1 means only 1 word is
considered for each step (token), resulting in deterministic completions,
while 40 means 40 words are considered at each step. 0 (default) is a
special setting meaning no restrictions. 40 generally is a good value.}
}
\description{
Sample from a model conditioning on input
}
