% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{gpt2_download}
\alias{gpt2_download}
\title{Download Model}
\usage{
gpt2_download(model = c("124M", "355M", "774M"))
}
\arguments{
\item{model}{The size of the model to download: \code{"124M"}, \code{"355M"} or
\code{"774M"}.}
}
\description{
Downloads the pretrained models by size.
}
\details{
Download and caching is performed using the \code{pins} package and creates
pins prefixed with \code{gpt2_}.
}
