% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_gpt2}
\alias{install_gpt2}
\title{Install Dependencies}
\usage{
install_gpt2(method = c("auto", "virtualenv", "conda"), conda = "auto",
  envname = NULL, ...)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds
a method that will work in the local environment. See \code{?tensorflow::install_tensorflow}.}

\item{conda}{Path to conda executable (or "auto" to find conda using the
\code{PATH} and other conventional install locations).}

\item{envname}{Name of Python environment to install within.}

\item{...}{Additional parameters.}
}
\description{
Installs the python packages the GPT-2 model requires.
}
\details{
This model also requires Python 3, Python 2 is unsupported.
}
